+++ sbg/ord_unidim_dense_set.hpp
@@ -75,7 +75,7 @@ class OrdUnidimDenseSet : public SetStrategy {
 
   // Traditional set operations ------------------------------------------------
 
-  unsigned int cardinal() const override;
+  size_t cardinal() const override;
   bool isEmpty() const override;
   MD_NAT minElem() const override;
   MD_NAT maxElem() const override;
